backend/requirements.txt
pandas>=2.2.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.6.0
loguru>=0.7.0
typer>=0.12.0
pyyaml>=6.0.0
xlwings>=0.30.0
pywin32>=306
pyinstaller>=6.0.0